remove http server implementation
