+++ micronaut-mcp-server-java-sdk/src/main/java/io/micronaut/mcp/server/McpHttpServerSupplier.java
@@ -13,12 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.mcp.server.stateless.transport;
+package io.micronaut.mcp.server;
+
+import io.micronaut.core.annotation.Internal;
 
 import java.util.function.Supplier;
 
 /**
  * Strategy interface for resolving a {@link McpHttpServer}.
  */
+@Internal
 public interface McpHttpServerSupplier extends Supplier<McpHttpServer> {
 }
