+++ micronaut-mcp-server-java-sdk/src/main/java/io/micronaut/mcp/server/stateless/McpStatelessServerFactory.java
@@ -16,20 +16,59 @@
 package io.micronaut.mcp.server.stateless;
 
 import io.micronaut.context.annotation.Factory;
-import io.micronaut.http.HttpRequest;
-import io.micronaut.mcp.server.stateless.transport.HttpServerMcpStatelessServerTransport;
+import io.micronaut.core.annotation.Nullable;
 import io.modelcontextprotocol.server.McpStatelessServerHandler;
-import io.modelcontextprotocol.server.McpTransportContextExtractor;
+import io.modelcontextprotocol.common.McpTransportContext;
+import io.modelcontextprotocol.spec.McpSchema;
 import io.modelcontextprotocol.spec.McpStatelessServerTransport;
 import jakarta.inject.Singleton;
+import reactor.core.publisher.Mono;
+
+import java.util.Objects;
 
 /**
  * A factory for {@link McpStatelessServerHandler} and {@link McpStatelessServerTransport}.
  */
 @Factory
 final class McpStatelessServerFactory {
+
+    @Nullable
+    private McpStatelessServerHandler serverHandler;
+
     @Singleton
-    HttpServerMcpStatelessServerTransport<HttpRequest<?>> serverHandler(McpTransportContextExtractor<HttpRequest<?>> extractor) {
-        return new HttpServerMcpStatelessServerTransport<>(extractor);
+    McpStatelessServerTransport serverHandler() {
+        return new McpStatelessServerTransport() {
+            @Override
+            public void setMcpHandler(McpStatelessServerHandler mcpHandler) {
+                serverHandler = mcpHandler;
+            }
+
+            @Override
+            public Mono<Void> closeGracefully() {
+                return Mono.empty();
+            }
+        };
     }
+
+    @Singleton
+    McpStatelessServerHandler serverTransport() {
+        return new McpStatelessServerHandler() {
+            @Override
+            public Mono<McpSchema.JSONRPCResponse> handleRequest(McpTransportContext transportContext, McpSchema.JSONRPCRequest request) {
+                checkServerHandler();
+                return serverHandler.handleRequest(transportContext, request);
+            }
+
+            @Override
+            public Mono<Void> handleNotification(McpTransportContext transportContext, McpSchema.JSONRPCNotification notification) {
+                checkServerHandler();
+                return serverHandler.handleNotification(transportContext, notification);
+            }
+
+            private void checkServerHandler() {
+                Objects.requireNonNull(serverHandler, "McpStatelessServerHandler is not set");
+            }
+        };
+    }
+
 }
