+++ micronaut-mcp-server-java-sdk/src/test/java/io/micronaut/mcp/server/stateless/McpControllerTest.java
@@ -1,7 +1,6 @@
-package io.micronaut.mcp.server.stateless.transport;
+package io.micronaut.mcp.server.stateless;
 
 import io.micronaut.http.HttpStatus;
-import io.micronaut.mcp.server.stateless.transport.HttpServerMcpStatelessServerTransport;
 import io.modelcontextprotocol.spec.McpError;
 import io.modelcontextprotocol.spec.McpSchema;
 import org.junit.jupiter.api.Test;
@@ -37,15 +36,15 @@ class McpControllerTest {
             .build()
             .getJsonRpcError();
 
-        int actual = HttpServerMcpStatelessServerTransport.status(jsonRpcError);
-        assertEquals(expected.getCode(), actual);
+        HttpStatus actual = McpController.status(jsonRpcError);
+        assertEquals(expected, actual);
     }
 
     @Test
     void errorJsonrpcResponse() {
         McpSchema.JSONRPCRequest request = new McpSchema.JSONRPCRequest("2.0", "prompts/list", 3, null);
         McpError error = new McpError("Missing handler request type: prompts/list");
-        McpSchema.JSONRPCResponse jsonrpcResponse = HttpServerMcpStatelessServerTransport.errorJsonrpcResponse(request, error);
+        McpSchema.JSONRPCResponse jsonrpcResponse = McpController.errorJsonrpcResponse(request, error);
         assertNotNull(jsonrpcResponse);
         assertNotNull(jsonrpcResponse.error());
     }
