+++ settings.gradle
@@ -20,8 +20,6 @@ include 'test-suite-jackson-databind'
 include 'test-suite-mcp-http-tck'
 include 'test-suite-mcp-http-tck-sync'
 include 'test-suite-mcp-http-tck-async'
-include 'test-suite-mcp-http-tck-httpserver-sync'
-include 'test-suite-mcp-http-tck-httpserver-async'
 
 enableFeaturePreview 'TYPESAFE_PROJECT_ACCESSORS'
