+++ test-suite-mcp-http-tck-async/src/test/resources/META-INF/services/io.micronaut.mcp.server.McpHttpServerSupplier
