--- test-suite-mcp-http-tck-httpserver-async/src/test/java/io/modelcontextprotocol/server/http/tck/httpserver/sync/HttpServerAsyncSuite.java
@@ -1,13 +1,0 @@
-package io.modelcontextprotocol.server.http.tck.httpserver.sync;
-
-import org.junit.platform.suite.api.SelectPackages;
-import org.junit.platform.suite.api.Suite;
-import org.junit.platform.suite.api.SuiteDisplayName;
-
-@SelectPackages({
-    "io.modelcontextprotocol.server.http.tck"
-})
-@Suite
-@SuiteDisplayName("MCP HTTP Server TCK for Java built-in HTTP Server Async")
-public class HttpServerAsyncSuite {
-}
