--- test-suite-mcp-http-tck-httpserver-async/src/test/resources/META-INF/services/io.micronaut.mcp.server.stateless.transport.McpHttpServerSupplier
@@ -1,1 +1,0 @@
-io.modelcontextprotocol.server.http.tck.httpserver.sync.AsyncMcpHttpServerSupplier
