--- test-suite-mcp-http-tck-httpserver-sync/src/test/resources/logback.xml
@@ -1,12 +1,0 @@
-<configuration>
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%cyan(%d{HH:mm:ss.SSS}) %gray([%thread]) %highlight(%-5level) %magenta(%logger{36}) - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <root level="info">
-        <appender-ref ref="STDOUT" />
-    </root>
-    <logger name="io.micronaut.http.client" level="TRACE"/>
-</configuration>
