+++ test-suite-mcp-http-tck-sync/src/test/resources/META-INF/services/io.micronaut.mcp.server.McpHttpServerSupplier
