+++ test-suite-mcp-http-tck/src/main/java/io/modelcontextprotocol/server/http/tck/HttpRequestUtils.java
@@ -15,7 +15,7 @@
  */
 package io.modelcontextprotocol.server.http.tck;
 
-import io.micronaut.mcp.server.stateless.transport.McpHttpServer;
+import io.micronaut.mcp.server.McpHttpServer;
 
 import java.net.URI;
 import java.net.URISyntaxException;
