+++ test-suite-mcp-http-tck/src/main/java/io/modelcontextprotocol/server/http/tck/PromptsGetTest.java
@@ -15,7 +15,7 @@
  */
 package io.modelcontextprotocol.server.http.tck;
 
-import io.micronaut.mcp.server.stateless.transport.McpHttpServer;
+import io.micronaut.mcp.server.McpHttpServer;
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
